* do not pipe pandoc output but use --log= option
Main problem was that pandoc changed the way it prints verbose output.
In old versions it was going to stdout but in newer versions to stderr.
Now we get a JSON log file with some of the info.
If more real verbose output is needed the --verbose option can still
be provided to PARAMS.

See: https://github.com/jgm/pandoc/issues/6628